Add JSON download functionality for trace details
